feat: mypage controller 추가

+++ src/main/java/com/lib/ailibrary/domain/room/RoomReserveResponse.java
@@ -18,6 +18,6 @@ public class RoomReserveResponse {
     private LocalDate rezDate;       // 예약 날짜
     private String rezTime;     // 예약 시간
     private int roomId;         // 시설번호
-    private int UserStuId;   // 예약자 학번
+    private Long UserStuId;   // 예약자 학번
     private String UserName;    // 예약자 성명
 }
